requirements.txt
Django==2.1.2
channels==2.1.3
dj_database_url==0.5.0
psycopg2==2.7.5
djangorestframework==3.8.2